Persist Scale settings Fixes #5
