+++ src/components/range_tool/ranges.jsx
